fix: save files in correct SWC format
All non-commented columns must be nodes in the neuron reconstruction. This function was saving the columns as the first row using pd.to_csv. However, this adds the column names as the first row uncommented which is incompatible with SWC specs. This commit creates the swc and prepends the column labels as a comment.