update (#7)
* refactor get_neuron_swc to process SWC data more efficiently

* version updates

* Refactor SWC module and add type hints

* add parallel processing for downloading swc files

* fix: save files in correct SWC format

All non-commented columns must be nodes in the neuron reconstruction. This function was saving the columns as the first row using pd.to_csv. However, this adds the column names as the first row uncommented which is incompatible with SWC specs. This commit creates the swc and prepends the column labels as a comment.

* use uv.lock file

* replace poetry with uv

* add requirements.txt

* update versions

* update recipes to use uv run

* change theme

* update  deps

* update hooks

* use updated uv commands

* update pyproject info

* update package versions

* update installation instructions